project with models and no db
