+++ src/models/index.js
@@ -1,0 +1,30 @@
+let users = {
+    1: {
+      id: '1',
+      username: 'Conrad P.B',
+    },
+    2: {
+      id: '2',
+      username: 'Tia reed',
+    },
+  };
+  
+let messages = {
+    1: {
+      id: '1',
+      text: 'Hello all, How u today?',
+      userId: '1',
+    },
+    2: {
+      id: '2',
+      text: 'Itsss all gooodd brahh!!',
+      userId: '2',
+    },
+  };
+
+  export default {
+    users,
+    messages,
+  }
+  
+
