+++ src/server.js
@@ -2,23 +2,32 @@ import { v4 as uuidv4 } from 'uuid'
 import 'dotenv/config'
 import express from 'express'
 import cors from 'cors'
+import models from './models'
 
 const app = express()
 
+
 app.use(cors())
-app.use(express.json());
+app.use(express.json())
 app.use(express.urlencoded({ extended: true }))
 app.use((req, res, next) => {
-  req.me = users[1]
+  req.context = {
+    models,
+    me: models.users[1],}
   next();
 })
 
+
+app.get('/session', (req, res) => {
+  return res.send(req.context.models.users[req.context.me.id]);
+})
+
 app.get('/users', (req, res) => {
-  return res.send(Object.values(users))
+  return res.send(Object.values(req.context.models.users))
   })
 
 app.get('/users/:userId', (req, res) => {
-  return res.send(users[req.params.userId])
+  return res.send(req.context.models.users[req.params.userId])
   })
 
 app.post('/users', (req, res) => {
@@ -38,11 +47,11 @@ app.delete('/users/:userId', (req, res) => {
   })
 
   app.get('/messages', (req, res) => {
-    return res.send(Object.values(messages))
+    return res.send(Object.values(req.context.models.messages))
     })
   
   app.get('/messages/:messageId', (req, res) => {
-    return res.send(messages[req.params.messageId])
+    return res.send(req.context.models.messages[req.params.messageId])
     })
 
   app.post('/messages', (req, res) => {
@@ -50,42 +59,37 @@ app.delete('/users/:userId', (req, res) => {
   const message = {
     id,
     text: req.body.text,
-    userId: req.me.id
+    userId: req.context.me.id,
+   
   };
 
-  messages[id] = message;
+  req.context.models.messages[id] = message;
 
   return res.send(message);
 })
   
-    
+app.put('/messages/:messageId', (req, res) => {
+  
+  return res.send(`PUT HTTP method on user/${req.params.userId} resource`)
+  
+})
 
-  let users = {
-    1: {
-      id: '1',
-      username: 'Conrad P.B',
-    },
-    2: {
-      id: '2',
-      username: 'Tia reed',
-    },
-  };
   
-  let messages = {
-    1: {
-      id: '1',
-      text: 'Hello all, How u today?',
-      userId: '1',
-    },
-    2: {
-      id: '2',
-      text: 'Itsss all gooodd brahh!!',
-      userId: '2',
-    },
-  };
+app.delete('/messages/:messageId', (req, res) => {
+  const {
+    [req.params.messageId] : message,
+    ...otherMessages
+  } = req.context.models.messages
+
+  req.context.models.messages = otherMessages
+
+  return res.send(
+    message
+    )
+  })
+
   
-  
-app.listen(process.env.PORT, () =>
+
+  app.listen(process.env.PORT, () =>
   console.log(`App is hot and listening on port ${process.env.PORT}!`),
 )
-
